refactor!: remove old AmplitudeModel
