+++ docs/adr/001.md
@@ -9,8 +9,8 @@
 ## Context and problem statement
 
 From the perspective of a PWA fitter package, the responsibility of the
-`expertsystem` is to construct an `AmplitudeModel` that serves as blueprint for
-a function that can be evaluated. Such a **function** has the following
+{mod}`expertsystem` is to construct a `AmplitudeModel` that serves as blueprint
+for a function that can be evaluated. Such a **function** has the following
 requirements:
 
 1. It should be able to compute a list of real-valued intensities
