+++ tests/channels/test_d0_to_ks_kp_km.py
@@ -1,7 +1,7 @@
 import expertsystem as es
 
 
-def test_script(output_dir):
+def test_script():
     result = es.generate_transitions(
         initial_state="D0",
         final_state=["K~0", "K+", "K-"],
@@ -21,5 +21,4 @@ def test_script(output_dir):
         "phi(1020)",
     }
     model = es.generate_amplitudes(result)
-    assert len(model.parameters) == 11
-    es.io.write(model, output_dir + "D0_to_K0bar_Kp_Km.yml")
+    assert len(model.parameters) == 5
