+++ tests/unit/conftest.py
@@ -4,7 +4,11 @@ import logging
 import pytest
 
 import expertsystem as es
-from expertsystem.amplitude.model import AmplitudeModel
+from expertsystem.amplitude.dynamics import set_resonance_dynamics
+from expertsystem.amplitude.dynamics.builder import (
+    create_relativistic_breit_wigner_with_ff,
+)
+from expertsystem.amplitude.helicity import HelicityModel
 from expertsystem.reaction import Result
 
 logging.basicConfig(level=logging.ERROR)
@@ -35,19 +39,21 @@ def jpsi_to_gamma_pi_pi_helicity_solutions() -> Result:
 @pytest.fixture(scope="session")
 def jpsi_to_gamma_pi_pi_canonical_amplitude_model(
     jpsi_to_gamma_pi_pi_canonical_solutions: Result,
-) -> AmplitudeModel:
+) -> HelicityModel:
     return __create_model(jpsi_to_gamma_pi_pi_canonical_solutions)
 
 
 @pytest.fixture(scope="session")
 def jpsi_to_gamma_pi_pi_helicity_amplitude_model(
     jpsi_to_gamma_pi_pi_helicity_solutions: Result,
-) -> AmplitudeModel:
+) -> HelicityModel:
     return __create_model(jpsi_to_gamma_pi_pi_helicity_solutions)
 
 
-def __create_model(result: Result) -> AmplitudeModel:
+def __create_model(result: Result) -> HelicityModel:
     model = es.amplitude.generate(result)
     for name in result.get_intermediate_particles().names:
-        model.dynamics.set_breit_wigner(name)
+        set_resonance_dynamics(
+            model, name, create_relativistic_breit_wigner_with_ff
+        )
     return model
